Fix dangling ellipsis state
If the same StaticLayout was used for different texts, the calculated
ellipsis state would not be reset for a second text that does not need
ellipsis, when the first text needed ellipsis. In practice, the text
would be incorrectly ellipsized.

This happens in practice when DynamicLayout uses StaticLayout, as it
keeps a StaticLayout around in a static member variable, and calls
generate() to reflow text, with varying text.

Change-Id: I866e59ce503c95054b355f10afd183fe7574dc81

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>